move math methods.
