+++ app/backend/app/Library/Math/MathLibrary.php
@@ -80,4 +80,180 @@ class MathLibrary
         // a **((p-1)(q-1)n + 1) ≡ a mod pq
         return ($value ** ($mod - 2)) % $mod;
     }
+
+    /**
+     * get prime factorization values (素因数分解)
+     *
+     * @param int $value value
+     * @return array
+     */
+    public static function getPrimeFactorization(int $value): array
+    {
+        $factors = [];
+        $divisor = 2;
+        $number = $value;
+
+        while ($number > 1) {
+            // 割り切れる場合
+            if ($number % $divisor === 0) {
+                // 素因数を格納
+                $factors[] = $divisor;
+                $number = $number / $divisor;
+            } else {
+                $divisor++;
+            }
+        }
+
+        return $factors;
+    }
+
+    /**
+     * get max prime factorization has two values (素因数分解)
+     *
+     * @param int $value value
+     * @return int
+     */
+    public static function getMaxTwoPairPrimeFactorization(int $value): int
+    {
+        $result = 0;
+        // 最大値の為パラメーターから減算して確認
+        for ($i = $value; 0 < $i; $i--) {
+            if ($i / 2 === 0) {
+                continue;
+            }
+            // 素因数分解の結果2つだけ取得出来る数値を取得
+            if (count(self::getPrimeFactorization($i)) === 2) {
+                $result = $i;
+                break;
+            }
+
+        }
+        return $result;
+    }
+
+    /**
+     * get two pair factorization values at $maxCount (素因数分解)
+     *
+     * @param int $value value
+     * @return int
+     */
+    public static function getMaxCoountTwoPairPrimeFactorization(int $value, int $maxCount): int
+    {
+        $result = 0;
+        $count = 0;
+        // 指定された値を最大値として2から$maxCount目の値を取得
+        for ($i = 2; 0 < $value; $i++) {
+            if ($i / 2 === 0) {
+                continue;
+            }
+            // 素因数分解の結果2つだけ取得出来る数値を取得
+            if (count(self::getPrimeFactorization($i)) === 2) {
+                $result = $i;
+                ++$count;
+                if ($count === $maxCount) {
+                    break;
+                }
+            }
+
+        }
+        return $result;
+    }
+
+    /**
+     * get greatest common divisor value (GCD=最大公約数)
+     *
+     * @param int $value1 value
+     * @param int $value2 compair value
+     * @return int
+     */
+    public static function getGreatestCommonDivisor(int $value1, int $value2): int
+    {
+        // 割り切れたら終了
+        while ($value2 != 0) {
+            $remainder = $value1 % $value2;
+            $value1 = $value2;
+            $value2 = $remainder;
+        }
+        // 絶対値に変換
+        return abs($value1);
+    }
+
+    /**
+     * check is greatest common divisor value is 1
+     *
+     * @param int $value1 value
+     * @param int $value2 compair value
+     * @return bool
+     */
+    public static function isGcdIsOne(int $value1, int $value2): bool
+    {
+        // 最大公約数が1=互いに素な値
+        return self::getGreatestCommonDivisor($value1, $value2) === 1;
+    }
+
+    /**
+     * get max greatest common divisor number in parameter($value以下の数字で$targetと互いに素の値の最大値を取得)
+     *
+     * @param int $value
+     * @param int $target
+     * @return int
+     */
+    public static function getMaxGreatestCommonDivisor(int $value, int $target): int
+    {
+        $result = 0;
+        // 最大値の為パラメーターから減算して確認
+        for ($i = $value; 0 < $i; $i--) {
+            if (self::isGcdIsOne($i, $target)) {
+                $result = $i;
+                break;
+            }
+
+        }
+        return $result;
+    }
+
+    /**
+     * get least common multiple value (LCM=最小公倍数)
+     *
+     * @param int $value1 value
+     * @param int $value2 compair value
+     * @return int
+     */
+    public static function getLeastCommonMultiple(int $value1, int $value2): int
+    {
+        $gcd = self::getGreatestCommonDivisor($value1, $value2);
+        return ($value1 * $value2) / $gcd;
+    }
+
+    /**
+     * get GCD of $value1, $value2, And x,y (拡張ユークリッド互除法)
+     * ユークリッドの互除法を用いてex + ly = gcd(e,l)の解となる整数x,yの組を見つける
+     *
+     * @param int $value1 value
+     * @param int $value2 compair value
+     * @return array
+     */
+    public static function getExtendedEuclidean(int $value1, int $value2): array
+    {
+        if ($value2 === 0) {
+            return ['gcd' => $value1, 'x' => 1, 'y' => 0];
+        }
+
+        // 商
+        $quotient = intval($value1 / $value2);
+        // 余り
+        $remainder = $value1 % $value2;
+
+        // 再起的に取得
+        $result = self::getExtendedEuclidean($value2, $remainder);
+        $gcd = $result['gcd'];
+        $x = $result['y'];
+        $y = $result['x'] - ($quotient * $result['y']);
+
+        return [
+            'gcd' => $gcd,
+            'x' => $x,
+            'y' => $y,
+        ];
+    }
 }
