+++ app/backend/tests/Unit/Library/Math/PrimeNumberLibraryTest.php
@@ -116,65 +116,6 @@ class PrimeNumberLibraryTest extends TestCase
     }
 
     /**
-     * max get prime factorization data
-     * @return array
-     */
-    public function getPrimeFactorizationDataProvider(): array
-    {
-        $this->createApplication();
-
-        return [
-            "value is 30" => [
-                'value' => 30,
-                'expect' => [2,3,5],
-            ],
-            "value is 33" => [
-                'value' => 33,
-                'expect' => [3,11],
-            ],
-            "value is 84" => [
-                'value' => 84,
-                'expect' => [2,2,3,7],
-            ],
-            "value is 100" => [
-                'value' => 100,
-                'expect' => [2,2,5,5],
-            ],
-            "value is 200" => [
-                'value' => 200,
-                'expect' => [2,2,2,5,5],
-            ],
-        ];
-    }
-
-    /**
-     * max get greated common divisor data
-     * @return array
-     */
-    public function getGreatestCommonDivisorDataProvider(): array
-    {
-        $this->createApplication();
-
-        return [
-            "value1=3/value2=7/expect=1" => [
-                'value1' => 3,
-                'value2' => 7,
-                'expect' => 1,
-            ],
-            "value1=5/value2=15/expect=5" => [
-                'value1' => 5,
-                'value2' => 15,
-                'expect' => 5,
-            ],
-            "value1=100/value2=24/expect=4" => [
-                'value1' => 100,
-                'value2' => 24,
-                'expect' => 4,
-            ],
-        ];
-    }
-
-    /**
      * test is prime number.
      *
      * @dataProvider isPrimeNumberDataProvider
@@ -212,32 +153,5 @@ class PrimeNumberLibraryTest extends TestCase
     public function testGetGreaterPrimeNumbers(int $value, int $count, array $expect): void
     {
         $this->assertSame($expect, PrimeNumberLibrary::getGreaterPrimeNumbers($value, $count));
-    }
-
-    /**
-     * test get prime factorization.
-     *
-     * @dataProvider getPrimeFactorizationDataProvider
-     * @param int $value
-     * @param array $expect
-     * @return void
-     */
-    public function testGetPrimeFactorization(int $value, array $expect): void
-    {
-        $this->assertSame($expect, PrimeNumberLibrary::getPrimeFactorization($value));
-    }
-
-    /**
-     * test get greatest common divisor.
-     *
-     * @dataProvider getGreatestCommonDivisorDataProvider
-     * @param int $value1
-     * @param int $value2
-     * @param int $expect
-     * @return void
-     */
-    public function testGetGreatestCommonDivisor(int $value1, int $value2, int $expect): void
-    {
-        $this->assertSame($expect, PrimeNumberLibrary::getGreatestCommonDivisor($value1, $value2));
     }
 }
